Improve tests (#9)
Better tests for model prompts, task evaluation failure handling and benchmarking assets.

# Detailed commit history:

* Implement dynamic types for data samples while testing model prompts

Previously, testing of model prompt functions assumed that an input
was of type `str`. This commit removes this incorrect assumption,
and fetches the required format from a newly defined `Dataset`
function.

* Update testing dependencies

* Add tests for task evaluation functions on prediction failure

The prediction pipeline may fail at several stages due to model,
pre-processing or post-processing errors; the prediction is set to
`None`, so that the `Task` may handle this failure as per its
semantics. This commit adds tests that ensure that such cases are
handled by the evaluation function.

* Add test for evaluate function output type

* Improve benchmark asset tests