Update model classes with generic names (#203)
This commit removes the obsolete RandomGPT model, renames GPTModel to LegacyOpenAIModel (since its based on the old completion API), GPTChatCompletionModel to OpenAIModel and BLOOMPetalModel to PetalsModel. This is done to reflect the generic nature of these models (i.e. PetalsModels can connect to any model hosted using Petals, not just BLOOM)

* Remove obsolete RandomGPT model

A better, more generic Random model will be introduced in the near future.
This commit removes the existing (limited) RandomGPT model and its references.
(The model was unused in all committed assets)

* Rename GPTModel to OpenAIModel

* Rename BLOOMPetalModel to PetalsModel