Implement graceful failure when an asset fails to run. (#206)
Currently, a failing asset will cause the benchmark to stop executing. This commit relaxes this behavior by logging the error for the misbehaving asset and continues to run other assets. The commit also adds some tests for the benchmark runner.

* Implement graceful failure when an asset fails to run.

Currently, a failing asset will cause the benchmark to stop executing.
This commit relaxes this behavior by logging the error for the misbehaving
asset and continues to run other assets.

* Add tests for benchmark runner

* Add test for multiconfig assets