Implement relative path handling in datasets and assets (#230)
All assets/datasets now use relative paths by default (with the option to fallback to absolute paths like before). All pre-defined paths are not relative, with a path resolver that automatically constructs the full absolute path. Also added the option to supply the base data path with `--data_dir` to the benchmark.

* Make 'data_dir' a mandatory argument of Dataset class

* Remove 'Dataset' suffix from downloaded data

* Rename 'Q2QSim' to 'STSQ2Q'

* Convert paths in assets to relative paths

* Replace all paths in datasets with relative paths, and add path resolver to all loaders

* Fix bug in :depend: based paths